# Add fast reductions