Fixes https://github.com/JuliaData/DataFrames.jl/issues/2439 https://github.com/JuliaData/DataFrames.jl/issues/2768 https://github.com/JuliaData/DataFrames.jl/issues/2440

I have implemented a proof of concept for `sum`. Let us discuss if we like this approach. If we do we can extend the list of supported reductions (and comments what reductions we would like to support are welcome).

CC @nalimilan @pdeffebach