@bkamins 2021-09-09
TODO: when we decide on what we do update documentation and NEWS.md