@bkamins 2021-09-10
@nalimilan The thing to discuss is the following tension:
This is what we currently do:
```
julia> df = DataFrame(x1=[1,2,missing], x2=[4.0,missing,missing])
3×2 DataFrame
 Row │ x1       x2        
     │ Int64?   Float64?  
─────┼────────────────────
   1 │       1        4.0
   2 │       2  missing   
   3 │ missing  missing   

julia> select(df, AsTable(:) => ByRow(sum∘skipmissing))
3×1 DataFrame
 Row │ x1_x2_sum_skipmissing 
     │ Real
─────┼───────────────────────
   1 │                   5.0
   2 │                   2
   3 │                   0

julia> df = DataFrame(x1=Any[missing,2,3], x2=Any[4.0,missing,missing])
3×2 DataFrame
 Row │ x1       x2      
     │ Any      Any     
─────┼──────────────────
   1 │ missing  4.0     
   2 │ 2        missing 
   3 │ 3        missing 

julia> select(df, AsTable(:) => ByRow(sum∘skipmissing))
3×1 DataFrame
 Row │ x1_x2_sum_skipmissing 
     │ Real
─────┼───────────────────────
   1 │                   4.0 
   2 │                   2   
   3 │                   3   
```
Which I think is not very nice. I would rather prefer to do `promote_type` of `eltype` of all columns and based on this determine the eltype of the output. In the first case above it would nicely produce a vector of `Float64` - which I think is preferable.
However, the consequence is that having vectors of `Any` would be problematic as we would not be able to establish `0` in this case. What I would propose to do is to internally narrow-down the eltype of such vectors. If it is still `Any` - error (as it means we have non numeric or missing values in the column). Otherwise perform the calculation using the columns with narrowed down eltype. What do you think?