@nalimilan 2021-09-10
Use `promote_type` would indeed make sense. That would be a difference between `ByRow` and plain broadcasting, right?

Regarding vector with eltype `Any`, I'd rather not make any particular efforts to make them work. Ideally we would do the same thing as without the fast path, even if that involves taking the slow path for them.