@bkamins 2021-09-10
> That would be a difference between `ByRow` and plain broadcasting, right?

It is unrelated to `ByRow`/broadcasting. Neither `map` nor broadcasting do `promote_type` the way I propose (as you see in my examples the problem is that each row is treated differently based on mixture of information about both eltypes of columns and values stored in a given row). What I propose is essentially to determine the result type based ONLY on eltypes of columns (which is superior to what what we have now except for the case of columns with `Any` type).

> Ideally we would do the same thing as without the fast path, even if that involves taking the slow path for them.

The problem is that taking a slow path will not yield any result as for wide inputs Julia crashes - this PR is meant to avoid these crashes in the first place.
Instead of narrowing of eltype we could also just error if the column has `Any` eltype and produce an informative error message. I think in general we should discourage users from using `Any` eltype in columns.