@bkamins 2021-09-11
I have pushed the proposal for summation with skipping missings. It works as follows:
```
julia> df = DataFrame(x1=[1,2,missing], x2=[4.0,missing,missing])
3×2 DataFrame
 Row │ x1       x2        
     │ Int64?   Float64?  
─────┼────────────────────
   1 │       1        4.0 
   2 │       2  missing   
   3 │ missing  missing   

julia> select(df, AsTable(:) => ByRow(sum∘skipmissing))
3×1 DataFrame
 Row │ x1_x2_sum_skipmissing 
     │ Float64
─────┼───────────────────────
   1 │                   5.0 
   2 │                   2.0 
   3 │                   0.0 

julia> df = DataFrame(x1=Any[missing,2,3], x2=Any[4.0,missing,missing])
3×2 DataFrame
 Row │ x1       x2      
     │ Any      Any     
─────┼──────────────────
   1 │ missing  4.0     
   2 │ 2        missing 
   3 │ 3        missing 

julia> select(df, AsTable(:) => ByRow(sum∘skipmissing))
ERROR: ArgumentError: The reduced element type Any does not support zero that is required to perform summation. Narrowing down element types of passed columns should be performed first.    
```

It is possible to try harder to deduce the summation result in the second case, but I am not sure it is useful (it would add a lot of complexity to the code in the case that I am not sure we want to support anyway).