@nalimilan 2021-09-11
I see two difficult points to address:
- Changing the behavior of `ByRow` is breaking. Minor changes to the eltype are probably fine (especially if we make them more narrow). But throwing an error in cases that don't is problematic, and that would be the case for `Any`: even though it crashes with a large number of columns, it works in other cases. That's why I think that falling back to the current slow path would make sense to avoid any regressions.
- The fast path for aggregations should be an implementation detail. `ByRow(f)` should be equivalent to `ByRow(x -> f(x))`, modulo minor differences like rounding errors or possibly the former working in cases where the latter fails. So if we use `promote_type` for reductions, we should probably use it for all functions  (even if that implies converting the vector returned by `map` manually).