@bkamins 2021-09-11
> That's why I think that falling back to the current slow path would make sense to avoid any regressions.

OK - I will change it this way

> The fast path for aggregations should be an implementation detail. `ByRow(f)` should be equivalent to `ByRow(x -> f(x))`

I do not think it is possible. This is exactly the same problem with as fast aggegations for `GroupedDataFrame`.

If I write `AsTable(some_cols) => ByRow(f)` I do not know how `f` would use the `NamedTuple` passed. E.g. it might rely on column names etc.

The other problem is that even if we did some way to signal that `f` is a reduction the general approach will be slower than what is proposed here as in e.g. `sum` or `sum∘skipmissing` we explicitly use the knowledge of the way the reduction should be performed to make it efficient.