@nalimilan 2021-09-12
> I do not think it is possible. This is exactly the same problem with as fast aggegations for `GroupedDataFrame`.
> 
> If I write `AsTable(some_cols) => ByRow(f)` I do not know how `f` would use the `NamedTuple` passed. E.g. it might rely on column names etc.
> 
> The other problem is that even if we did some way to signal that `f` is a reduction the general approach will be slower than what is proposed here as in e.g. `sum` or `sum∘skipmissing` we explicitly use the knowledge of the way the reduction should be performed to make it efficient.

What do you mean with "not possible"? What I suggest is to only enable optimizations for known aggregation functions for which we know the equivalence holds -- just like for grouped reduction. Isn't that possible here, or do you just mean that it's not possible in general for any function?