@bkamins 2021-09-12
Ah, you mean with:

> `ByRow(f)` should be equivalent to `ByRow(x -> f(x))`

That the output should be identical - not the code path. Then it is possible but it will be slow and I am not sure if it is desirable. See the first example:
```
julia> df = DataFrame(x1=[1,2,missing], x2=[4.0,missing,missing])
3×2 DataFrame
 Row │ x1       x2        
     │ Int64?   Float64?  
─────┼────────────────────
   1 │       1        4.0
   2 │       2  missing   
   3 │ missing  missing   

julia> select(df, AsTable(:) => ByRow(sum∘skipmissing))
3×1 DataFrame
 Row │ x1_x2_sum_skipmissing 
     │ Real
─────┼───────────────────────
   1 │                   5.0
   2 │                   2
   3 │                   0
```
Retaining this behavior is possible, but will be complex to achieve and I would argue that the result is not something I think most people would expect or want.