@bkamins 2021-09-16
> Is adding a separate `table_transformation` step really less work than adding a `AsVector` wrapper?

The issue is two-fold:
1. performance - as you can see in the codes if we do e.g. `ByRow(sum∘skipmissing)` we actually do not process it by-row, but in a columnar fashion. This has the following consequence:

This PR:
```
julia> df = DataFrame(ones(10^7, 100), :auto);

julia> @time select(df, AsTable(:) => ByRow(sum∘skipmissing));
  2.946524 seconds (3.20 M allocations: 268.327 MiB, 14.51% gc time, 50.74% compilation time)

julia> @time select(df, AsTable(:) => ByRow(sum∘skipmissing));
  1.031689 seconds (255 allocations: 76.314 MiB, 0.59% gc time)

julia> @time select(df, AsTable(:) => ByRow(sum∘skipmissing));
  1.034903 seconds (255 allocations: 76.314 MiB, 0.34% gc time)
```
Current `main`:
```
julia> df = DataFrame(ones(10^7, 100), :auto);

julia> @time select(df, AsTable(:) => ByRow(sum∘skipmissing));
  5.808807 seconds (2.30 M allocations: 217.624 MiB, 6.73% gc time, 36.07% compilation time)

julia> @time select(df, AsTable(:) => ByRow(sum∘skipmissing));
  3.735678 seconds (365 allocations: 76.322 MiB, 0.07% gc time)

julia> @time select(df, AsTable(:) => ByRow(sum∘skipmissing));
  3.523625 seconds (365 allocations: 76.322 MiB, 0.05% gc time)
```
So as you can see it is over 3x slower because it processes data in a row-wise not whole column fashion.

So for performance we need custom code paths.

2. Correctness. `AsVector` would still require a custom code path for e.g. `ByRow(sum∘skipmissing)` to create an `init` value of proper type. And each reduction probably requires a different code for `init` value.

---

Having said that I for now opted-out from `AsVector` as we can add optimizations to `AsTable` path without complicating the minilanguage. Indeed the benefit of `AsVector` would be that users would be able to write "relatively fast" reductions that we did not think of (i.e. that are not hard coded). The question is how often do you think such a need arises? This is a discussion that would be good to have here before we finalize this PR.